Initial change
made the navbar be responsive according to JS, and recoded it with mobile-first in mind. Redesigned the index page banner, to not include logo inside of it, goes for all screensizes. Increased navbar min-width for 'desktop-view'. Next up is moving the navbar changes onto all pages & then adding form validation & add to cart UX